1. 누락 패치내역 추가.. 2. R25 중간작업.. 3. 번역관련 문제해결

+++ ElectronicObserver/Observer/kcsapi/api_req_kaisou/powerup.cs
@@ -11,7 +11,6 @@ namespace ElectronicObserver.Observer.kcsapi.api_req_kaisou
 	public class powerup : APIBase
 	{
 
-
 		public override void OnRequestReceived(Dictionary<string, string> data)
 		{
 
@@ -20,21 +19,23 @@ namespace ElectronicObserver.Observer.kcsapi.api_req_kaisou
 
 			db.Fleet.LoadFromRequest(this.APIName, data);
 
+            bool destroysEquipments = data.ContainsKey("api_slot_dest_flag") ? (data["api_slot_dest_flag"] != "0") : true;
 
-			foreach (string id in data["api_id_items"].Split(",".ToCharArray()))
-			{
+            foreach (string id in data["api_id_items"].Split(",".ToCharArray()))
+            {
+                int shipID = int.Parse(id);
 
-				int shipID = int.Parse(id);
+                ShipData ship = db.Ships[shipID];
+                if (destroysEquipments)
+                {
+                    for (int i = 0; i < ship.Slot.Count; i++)
+                    {
+                        if (ship.Slot[i] != -1)
+                            db.Equipments.Remove(ship.Slot[i]);
+                    }
+                }
 
-
-				ShipData ship = db.Ships[shipID];
-				for (int i = 0; i < ship.Slot.Count; i++)
-				{
-					if (ship.Slot[i] != -1)
-						db.Equipments.Remove(ship.Slot[i]);
-				}
-
-				Utility.Logger.Add(2, ship.NameWithLevel + " 를 제적했습니다.");
+                Utility.Logger.Add(2, ship.NameWithLevel + " 를 제적했습니다.");
 				db.Ships.Remove(shipID);
 
 			}
